+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Notes
 - **Add a "Quick Paste to Previous Text Field" option to the General settings**
   - When this setting is used, when you copy, copy as plain text, or change and copy from the History or Standard Phrase window, the focus is returned to the previously focused app before executing Quick Paste. If you are focused on a text field, you can easily copy and paste into the previous text field.
   - You can temporarily disable Quick Paste by holding down the `⌥ (Option)` key.
+- **Add a feature to recalculate the total size of the saved folder in the Copy History settings**
+  - The size of all files and folders included in the saved folder is now recalculated to display the correct size.
 
 ### Bug Fixes and Improvements
 - **Fix issue where copying a history with extremely long text from the History window could cause freezes or increased memory usage**